Finalized the implementation for assert.
It now returns a <FACT-n> symbol when you assert something. I'll
see if I can bind that symbol to the fact instance so that, at
least, it will make sense as a symbol and not just a printed
value.

I also modified the test to prep *fact-index* and *working-memory*
in the cleared state. I'll replace that with a call to (clear)
later on.