Added an implementation for binding the <FACT-n> symbols to fact instances.
This is not part of CLIPS but an addition I've chosen to implement
because I think assert should return the values that were asserted
not a symbol that indicates which index it has in working memory.
My half-ass middle ground is to bind the useless symbol as a defvar
to the fact instance which, sort of, achieves the same thing.